third_party/WebKit/LayoutTests/fast/html/imports/import-in-iframe-expected.txt

PASS wrongImport is null
PASS headlineText is "Hello"
PASS successfullyParsed is true

TEST COMPLETE
